gateway/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
httpx>=0.27
PyJWT[crypto]>=2.9
authlib>=1.3
google-cloud-firestore>=2.19
stripe>=11.0
pydantic>=2.9
pydantic-settings>=2.6
python-dotenv>=1.0
